Only post updated items to cart, not all items
This change means that for example if we have 30 items in the cart and we click "add to cart" on one more, we only send the data fo that one item and not for *all* 30 items that are in the cart. This means on the backend we then are doing one update instead of 30.