Remove cache logger initializer
I'm pretty sure this stopped working around Rails 5.2 (cache logs are no longer being sent to the Rails logger), and it's now causing a bug in Staging (with Redis). There have been a few changes in Rails around the way the cache store is instrumented and we need to reassess it a bit if we want to actually show caching info in the log output in dev and staging.